Bug 19240 (bad image list performance regression): * Don't connect to the commons DB on cache hits, in order to determine the cache key name. Removed remnants of that bright idea from GlobalFunctions.php. * Fixed total failure of negative caching in checkRedirect() due to memcached stripping trailing spaces from string values. Probably never worked.
Also:
* Respect hasSharedCache in foreign repos. Recently-added code was apparently ignorant of this setting.
* Renamed getMemcKey() to getSharedCacheKey() to make its function more clear, introduced getLocalCacheKey() to do the other thing. Fixed its parameters to be like wfMemcKey() and used it in more places.
* Used getLocalCacheKey() in various places instead of wfMemc(), to avoid having multiple repositories overwrite each others' caches. 
* Fixed the BagOStuff bug that the FIXME in LocalRepo::checkRedirect() appears to refer to.
* Removed getMasterDB() and getSlaveDB() from FileRepo, it's incorrect to assume that a repo other than a LocalRepo has an associated database.
* Made FileRepo::invalidateImageRedirect() a stub, to match FileRepo::checkRedirect(). Moved the functionality to LocalRepo where checkRedirect() is concretely implemented.

+++ includes/filerepo/LocalFile.php
@@ -133,11 +133,12 @@ class LocalFile extends File
 	}
 
 	/**
-	 * Get the memcached key
+	 * Get the memcached key for the main data for this file, or false if 
+	 * there is no access to the shared cache.
 	 */
 	function getCacheKey() {
 		$hashedName = md5($this->getName());
-		return wfMemcKey( 'file', $hashedName );
+		return $this->repo->getSharedCacheKey( 'file', $hashedName );
 	}
 
 	/**
@@ -590,8 +591,10 @@ class LocalFile extends File
 	function purgeHistory() {
 		global $wgMemc;
 		$hashedName = md5($this->getName());
-		$oldKey = wfMemcKey( 'oldfile', $hashedName );
-		$wgMemc->delete( $oldKey );
+		$oldKey = $this->getSharedCacheKey( 'oldfile', $hashedName );
+		if ( $oldKey ) {
+			$wgMemc->delete( $oldKey );
+		}
 	}
 
 	/**
